vlm_retrieval/requirements.txt
tqdm
numpy
pandas
beautifulsoup4
selenium
yt-dlp
requests

transformers
accelerate
pillow
torchvision